Border color change works!
Needed to call e.preventDefault() when nameValidator failed in order to see the results. Explanation from Sradms0:
"Your code actually works as it should :slightly_smiling_face:
However, you just can't see the color changes because your form submits, leaving the page.
In your forms 'submit listener you're calling your nameValidator function, but whether it catches something, or not, the form submits, and is too quick to see anything.
Note Robert's instruction in this listener:
  // 1. Create an if statement
    // If `(!nameValidator())` call `e.preventDefault();`
      // And log out a message saying this validator prevented submission
You want to throw your nameValidator call in an if-statement, checking if it catches invalid input in the name-field.
And if it does, you want to stop the form from submitting with e.preventDefault.
Once you've done this, you'll see the name-field borderColor change in accordance to invalid to valid input."