Still working on frameworkValidator()
There must be a simpler way to get the default selected value of the framework select. I created a function (line 83) to loop through all the options and return the .value of the option.defaultSelected, but it doesn't work. I feel like this should be accomplishable via a single line.